Use AnimatedCounterWidget for the changes pill above chat input
Extract AnimatedCounterWidget into a shared workbench/browser module so both
the changes summary widget and the chat turn changes pill can animate their
+insertions / -deletions counters.
