+++ src/vs/sessions/contrib/changes/browser/media/changesSummaryWidget.css
@@ -7,20 +7,6 @@
 	display: inline-flex;
 	gap: 4px;
 
-	.changes-summary-lines-added,
-	.changes-summary-lines-removed {
-		position: relative;
-		display: inline-grid;
-		overflow: hidden;
-		justify-items: start;
-
-		div {
-			grid-area: 1 / 1;
-			display: block;
-			will-change: transform, opacity;
-		}
-	}
-
 	.changes-summary-lines-added {
 		color: var(--vscode-chat-linesAddedForeground);
 	}
