+++ src/vs/workbench/browser/media/animatedCounterWidget.css
@@ -1,0 +1,21 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+/* Structural styles for the AnimatedCounterWidget. The stacked incoming/outgoing
+	value elements share a single grid cell so they can cross-fade and slide while
+	the container width tweens. Callers add their own class for color/typography. */
+
+.monaco-animated-counter {
+	position: relative;
+	display: inline-grid;
+	overflow: hidden;
+	justify-items: start;
+}
+
+.monaco-animated-counter > div {
+	grid-area: 1 / 1;
+	display: block;
+	will-change: transform, opacity;
+}
